This work is analog to what has been done for Thales Luna.
This playbook will configure the Barbican pods on the test system to use a Proteccio HSM as a crypto backend to store and generate keys.

The involved steps are:

1. Create modified barbican-api and barbican-worker images that contain the HSM client software. The new images will be published locally on the CRC node with a special tag ("cifmw_update_barbican_custom_tag") appended.
2. Create a secret to store certificates to access the HSM (server certificate, client certificate and key).
3. Create a secret to store the password needed to access the HSM device.
4. Use the update-containers role to modify openstackversion to use the updated Barbican images.
5. Modify the control plane CR to add the needed config to Barbican to use the Proteccio HSM as a backend.

Steps 1-3 are done by a separate Ansible role (https://github.com/openstack-k8s-operators/ansible-role-rhoso-proteccio-hsm/). This is useful because we'll be able to modify and branch this role as appropriate as the HSM software changes.

Jira: https://issues.redhat.com/browse/OSPRH-14750

Signed-off-by: Mauricio Harley <mharley@redhat.com>