Remove printing redhat-release in adoption_osp_deploy
The task should iterate on list of hosts, which should be provided
later as "_vm" variable. That was missing, so the Ansible was
failing because of undefined variable.
Printing the redhat-release should be added into pre-run config after
registration, but in this kind of CI jobs we did not spotted any issues
now.

Signed-off-by: Daniel Pawlik <dpawlik@redhat.com>