Make worker-Nodes take preference in nncp ci_kustomize
As for example in 3+3 scenarios, we want to run all openstack namespaces pods
in the worker nodes.

As we have different type of nodes: ocp, crc and ocp-worker; we want to
select the nodes by taking precedende to worker nodes if defined.
If not, we want to check if it's a SNO crc deployment, and lastly, we'll
go by typical scenarios where the control-plane, master and workers in the same
nodes.