rsync only if source directory exists
In reproducer role's configure_controller.yml tasks file,
stat the cifmw_reproducer_src_dir before attempting to
rsync it and then only rsync it if it exists. This is
a reasonable thing to do before calling rsync.

If the cifmw_reproducer_src_dir does not exist, then
fall back to github for the Ansible dependencies.
This will not break the current behavior for which
the rsync task was added.

Jira: OSPRH-17434

Signed-off-by: John Fulton <fulton@redhat.com>