adoption_osp_deploy: allow to pass private overcloud args
A new parameter allow users to specify an additional
overcloud deploy environment file (-e), which can be used
to pass private/restricted parameters.

cifmw_adoption_osp_deploy_overcloud_extra_args: points to a string
which contains an environment file.
The string is set to an empty value, so in the worst case the file
will just be empty, which is fine for `openstack overcloud deploy`.