Let sync workflow work with non-default branches
Currently the sync workflow checkouts always the default
branch in repository, so in case where we want to sync
other branch, such as `18.0-fr2`, the attempt of rebase
will result in unknown branch (as it was not pulled
by default to the locally cloned repository).

Hence, after this commit is applied, we will perform rebase
to what was explicitly fetched.