Remove test case name from class name
The validations framework currently puts test case name
as part of the class name field, which in the end means
that the test case name is duplicated,
e.g. as it is processed by Polarion or other JUnit parser.

This change alters the class name field to be just `validations`.