Fix terminology: use "architecture-based" instead of "VA"
The architecture repository contains severals DTs, some of which
are VAs. At the beginning the architecture-runner code was developed
with VA1 (later HCI VA) in mind, but we are really talking about
architecture-based deployments, not specifically to VAs.

This change only affects the ansible output and it has no functional
impact.