Move Ceph playbook into hooks; symlink ceph play to old location
We would like to avoid tasks, where we do 'import_playbook'.
In that case, instead of split ceph playbook into role (now),
let's keep the playbook as it is (because it is used in few places)
and run that play using hooks, like we do with other additional
services.
Until migration process is not done (many projects are using that
playbook), there is a symlink done to old location, so we should avoid
outage.

Related-to: https://github.com/openstack-k8s-operators/install_yamls/pull/1083
Related-to: https://github.com/openstack-k8s-operators/architecture/pull/604

Signed-off-by: Daniel Pawlik <dpawlik@redhat.com>