Call tasks from roles if play host is same
After we migrate some playbooks into roles, we can drop
the convention how it was done earlier - few plays where each
was done on same host and each was calling one (or few) task in role.
It is not needed, we can call other tasks in role in one play,
because all is done on same host.

Signed-off-by: Daniel Pawlik <dpawlik@redhat.com>