Avoid switching OpenShift project in hook
I encountered a case where the hook related to cinder
multi-attach volume type fails, as on some race condition
between commands execution the default OpenShift project
was changed and the assumed pod was not found.

To make the shell call more predictable, it is better
to run `oc` command with `-n` parameter, rather than
relying on switching the default project in earlier command.