Add required-projects to cifmw-pod-base job
In the run playbook `ci/playbooks/pod-jobs.yml` there is
a hard-coded dependency on CI-Framework in Zuul projects.
However, if that job is called in other repositories,
such as `edpm-ansible`, the ci-framework will not exist
in the Zuul environment. This change ensures the `ci-framework`
repository to be always present in job environment.