Do not fail if secret is not on remote host
Add ignore_errors=true to task which checks if the
remote secret is on the remote host so that the
play will not fail if `stat` gets permission denied.
Default the remote_src to false if the previous
task does not set _ps_exists as expected.

This is a follow to PR 3116, which helps to run
reproducer playbook locally and not on remote
host directly.

Jira: https://issues.redhat.com/browse/OSPRH-18840

Signed-off-by: John Fulton <fulton@redhat.com>