Document test-operator key file parameter
Recently, there was a problem with test-operator failing and the
debugging exposed that it was because test-operator searches for
the cifmw private key locally. Which means if it's being run on
a remote host, it fails. This patch adds a warning in form of
documentation to help users that might encounter this issue in
the future.