[test-operator][tempest] Pass the smoke param to allow running smoketests
The Tempest CR supports passing a parameter to run the smoketests.
This support is missing in the test-operator role, so the smoke tests
need to be enabled with a regex in the
cifmw_test_operator_tempest_include_list var, which also includes the
other tests to be run.

Tempest on it's own allows passing the --smoke argument. Adding a
boolean to control this in tests-operator role make the experience more
consistent across tempest, Tempest CR and the test-operator role.

This change simplifies the addition of smoketests, allowing just the
tests you're interested in to be specified, and the smoketests to be
toggled with a boolean, instead of explicitly adding an extra regex.

Murano smoke tests are disabled by default, since we don't have it in RHOSO.