Add retries to more kustomize_deploy tasks
We still observe the Internal Server Errors randomly
occurring from the underlying OpenShift cluster.
This commit extends the original attempt [1] that made
the situation a little bit better in some places,
such as when there is attempt to call webhook
"nodenetworkconfigurationpolicies-mutate.nmstate.io".

So, TL;DR we want to guard few more sections with retries,
to give the cluster some more time to stabilize and reply
with expected value.

[1] https://github.com/openstack-k8s-operators/ci-framework/pull/3068