Run nested ansible command and call run_logs tasks to drop 99-logs play
Earlier the 99-logs.yml playbook was executed as nested Ansible.
After we migrate that playbook to role, it can not be called directly,
because it contains: cimfw collection to be installed on Zuul executor
and include_vars can not read files, that are located on the controller,
so they needs to be synchronized with Zuul executor job dir.
Also fix condition that was changed when moving 99-logs.yml play into
the role.

Signed-off-by: Daniel Pawlik <dpawlik@redhat.com>