Add roles for fixing Python encodings
On the freshly reprovisioned nodes in our CI, not very often,
but still from time to time, we seem to lack the necessary
encoding file to invoke Ansible modules properly. I suspect
that all this happens due to some rogue cleaning script
that removes too many files, such as the compiled Python codes
(`*.pyc` files) and hence from time to time we hit this error
(the distro images we use are being rebuild quite often).

By including the new role as first task in the CI jobs,
we should never ever encounter this problem again.