Fix permission denied when checking for cifmw_reproducer_src_dir
The ansible stat module does gracefully handle permission denied, and
will instead fail with an error of the patch can't be read due to
permissions. Add an ignore_errors: true to the task checking for
cifmw_reproducer_src_dir and then verify the stat result is defined
before using it.

This scenario can occur when ansible_user is not the same user as the
one executing ansible-playbook, and since this task is delegated to
localhost, most users don't have permission to read other user's home
dirs.

Jira: OSPRH-19535
Signed-off-by: James Slagle <jslagle@redhat.com>