Check if hook is properly defined for execution
Adding the additional `when` conditions make it possible
to override the named hooks in job definitions with empty
mappings (i.e. `{}`) in order to skip specific ones.

Unfortunately, because `set_fact` tasks are involved
we cannot simply use the `undef()` to achieve that
(it raises `Mandatory variable has not been overridden.` [1]).

[1] https://stackoverflow.com/questions/73427379/how-to-unset-a-variable-in-ansible#comment138144412_78351300