Add pre and post logs hooks
This commit adds an option to define pre and post hooks
for the logs gathering phase – it will allow us to define
extra actions to be performed at the end of each job
no matter what was the last stage and status.

[PCP] Add role for managing Performance Co-Pilot

This role manages Performance Co-Pilot (PCP) toolkit
for monitoring and analyzing the historical details
of system performance. It would be useful for tracking
the resources utilization in the end-to-end testing jobs.

[PCP] Add custom pmrep metric specifications

[PCP] Add solution to plot metrics

This commit adds a script to produce figures from the collected
metrics files, using matplotlib and pandas libraries. Additionally
the corresponding Ansible tasks to invoke the script are also
provided in the `pcp_metrics` role.

[PCP] Make plot script compatible with Python 3.9

The `type | None` notation is possible from Python 3.10;
for Python 3.9 it is necessary to either use `Optional[]`
or `Union[]` specifier in typing.

[PCP] Add hooks to enable the metrics collection

[PCP] Annotations

[experimental] Figure auto-width