Drop unused playbooks except 08-run-tests.yml
The playbooks were migrated into role cifmw_setup.
As it was mentioned in commit message [1], such approach
should be better for developers to maintain the code.
Also skip removing 08-run-tests.yml because the NFV team is using it
and it takes a time to verify that. The playbook would be removed in
next pull request.

[1] https://github.com/openstack-k8s-operators/ci-framework/pull/2930

Depends-On: https://review.rdoproject.org/r/c/rdo-jobs/+/58147
Depends-On: https://github.com/openstack-k8s-operators/ci-framework/pull/3208
Signed-off-by: Daniel Pawlik <dpawlik@redhat.com>