DNM Checking reading group_vars until Zuul is not upgraded
We can not set in zuul job definition path to the dir or file [1]
that should be readded before Zuul executor starts the CI.
It makes troubles on "merging" common vars into the group_vars/all.
This PR shows if we can read the vars using cifmw_helpers role
to make it work.

[1] https://zuul-ci.org/docs/zuul/latest/config/job.html#attr-job.include-vars.name

Signed-off-by: Daniel Pawlik <dpawlik@redhat.com>