Treat various files vars as template
The Ansible set fact does not lookup to resolve the name if it
got a jinja2 template. It is an issue that in some cases, where the
variable files are like a template, for example:

    cifmw_repo_setup_basedir: "{{ cifmw_basedir | default(ansible_user_dir ~ '/ci-framework') }}"

Instead of resolving the name correctly, value is:

    {{ ansible_user_dir }}/ci-framework-data

It is an error, due later the modules are not "resolving" the name, but
raise an error:

    fatal: [localhost]: FAILED! =>
        changed: false
        cmd: /usr/bin/git ls-remote '{{' ansible_user_dir '}}/src/github.com/openstack-k8s-operators/repo-setup'
          -h refs/heads/HEAD
        msg: |-
          fatal: '{{' does not appear to be a git repository
          fatal: Could not read from remote repository.

          Please make sure you have the correct access rights
          and the repository exists.
        rc: 128

Parse the files as a template first, then read and set as fact.

Signed-off-by: Daniel Pawlik <dpawlik@redhat.com>