update: Add optional cleanup of OpenStack resources
This commit introduces tasks that cleans up resources created on
OpenStack during the update test. The cleanup runs by default.

A new boolean variable, `cifmw_update_dont_cleanup`, is added to allow
skipping the cleanup phase. This is useful for debugging purposes when
the created resources need to be inspected.

Closes: [openstack resources created to test update in rhoso 18 are not cleanup](https://issues.redhat.com/browse/OSPRH-20307)