WIP: Account for different users in reproducer
Using vars like ansilbe_user_id or ansible_user_dir in a task that also
uses `delegate_to: ...` can cause some issues.

For example, let's say we are running the reproducer from a machine with
the user 'exampleuser' and we have the user 'controlleruser' on controller-0.
If we have some task that is delegated to controller-0 and
tries to create some file in the controller-0 user's home directory, we
can't simply do:

```
- name: Create some file
  delegate_to: controller-0
  ansible.builtin.file:
    path: "{{ ansible_user_dir }}/some_file.txt"
    state: touch
```

`ansible_user_dir` is `/home/exampleuser`, and this path does not
exist on controller-0.

This has only worked because of the assumption that all hosts have a
'zuul' user.