Add building Tempest timing data url
There is a relatively new parameter added to test-operator, called
TimingDataUrl. We are trying to make the parameter work with jobs,
but the issue is that there is no variable that stores the final
Tempest pod name that will be created. Without it, we cannot build
the timing data url, as we are saving the data under job/test-name
folder.

This change adds a way to build the final url for both tests with
and without workflow.