Revert change to make preference nncp on worker nodes thatn master nodes
It seems these scenarios has master nodes and a single worker node which is tainted.
This worker node is used for running tempest but nothing related with OCP.

Problematic scenarios are those which has a worker node tainted for tempest, but the workload
is set in master nodes.

There's no reason to configure NNCP in master nodes when we have regular worker nodes, so we
were assuming that in case of there's a worker node, then all NNCP goes to worker node, which is
problematic when single worker node is tainted.

So we need to enhance the templates to:
* Check if there's a worker node, if so, check if those are tainted.
* If not tainted, worker takes preference.
* If tainted, master takes preference.
* If there's no worker node, then we go with master nodes.
* Same for SNO CRC scenarios.

Adding also uni02beta that takes the template from common