feat(reproducer): Allow non-zuul controller-0 user
It was already technically possible to create a controller-0 vm with a
non-zuul user. This commit enables configuring controller-0 with a
non-zuul user.

Also, this commit fixes instances where incorrect
user/user_dir variables were used on controller-0:
We can not use vars like ansible_user_id or ansible_user_dir with tasks
that are delgated to controller-0. If the host we are running the
reproducer from has a different user from controller-0, then vars like
`ansible_user_id` and `ansible_user_dir` will try to use the user from
the machine that is running the reproducer rather than controller-0's
user.

For example, let's say we are running the reproducer from a machine with
the user 'exampleuser' and we have the user 'controlleruser' on controller-0.
If we have some task that is delegated to controller-0 and
tries to create some file in the controller-0 user's home directory, we
can't simply do:

```
- name: Create some file
  delegate_to: controller-0
  ansible.builtin.file:
    path: "{{ ansible_user_dir }}/some_file.txt"
    state: touch
```

`ansible_user_dir` is `/home/exampleuser`, and this path does not
exist on controller-0.