fix(reproducer): Use controller-0 path when installing requirements
Previously, the check for `common-requirements.txt` would check if
common requirements exists on localhost at a given path. This path is
based on the home directory on localhost.
If the file did exist, it would use that same path to try to install the
requirements on controller-0. This causes issues when localhost and
controller-0 do not have the same users and home directories.