[operator_build] Update the go.mod replace to use local repos
When building from a PR that has multiple dependencies in the same repo,
the incorrect versions are built.
This happens because the build tasks extract a sha from an operators
list that is built from zuul.items, which contains a list of the
dependencies on a PR (from Depends-On). This list can contain multiple
PRs from the same repo, however, the operator list only takes the first
item for each repo.
The information is used to update the openstack-operator go.mod file
with a "replaces" directive that tell it where to find the operators to
deploy.
The line being added was referencing a particular commit on the PR
source branch previously.
This is incorrect when using zuul, as zuul uses speculative merging and
prepares a version of the repositories that have all the dependencies
merged together.
The "replaces" is updated to reference the local version of the
operator i.e. the one prepared by zuul.