adoption: allow files to be copied to the undercloud
Add a new (optional) key `additional_files` for each item in the
list of stacks.
`additional_files` is a list and each item is the name of a file
which must exist inside the scenario directory. The file will
be copied inside the home directory of the ansible user
(by default the `zuul` user) of the undercloud, so that
it can be used during the overcloud deployment.