Add Github action that verify commit message prefix
If the PR is related to some role, it would verify if in the title
or commit message header there is prefix set. For example, if
change was done for roles/cifmw_helpers, it should verify if the
title or commit message header contains [cifmw_helpers].

You can use a dedicated script: scripts/check-role-prefix.sh
to verify the commit locally.

Co-Authored-By: Szymon Datko <sdatko@redhat.com>
Signed-off-by: Daniel Pawlik <dpawlik@redhat.com>