Fix ceph 9 RGW deployment
rgw_frontend_ssl_certificate has been deprecated in ceph8 and in ceph9 doesn't
work properly anymore. There's a new way of setting both cert and key when ssl
is used and is fully documented in [1].
This patch still preserves the old way of deploying rgw through a new var used
to execute the old code. When "rgw_ssl_backward_compatibility" is set,
the old facts are set, resulting in populating the old variables, otherwise
the new method based on ssl_cert and ssl_key is applied.

[1] https://docs.ceph.com/en/latest/cephadm/services/rgw/

Signed-off-by: Francesco Pantano <fpantano@redhat.com>