Set timeout in each oc adm must-gather command execution
It happens that the "oc adm must-gather" command takes longer
than expected, so the CI job gets timeout and the collecting
logs process is disturbed. In some places, we spotted that the
'--timeout' parameter in 'oc adm' command is just ignored:

    TASK [os_must_gather : Run openstack-must-gather command
        output_dir={{ cifmw_os_must_gather_output_dir }}/artifacts,
        script=oc adm must-gather --image {{ cifmw_os_must_gather_image }}
        --timeout {{ cifmw_os_must_gather_timeout }}
        --host-network={{ cifmw_os_must_gather_host_network }}
        --dest-dir {{ cifmw_os_must_gather_output_log_dir }}
        -- ADDITIONAL_NAMESPACES={{ cifmw_os_must_gather_additional_namespaces }}
           OPENSTACK_DATABASES=$OPENSTACK_DATABASES SOS_EDPM=$SOS_EDPM SOS_DECOMPRESS=$SOS_DECOMPRESS gather 2>&1]
    POST-RUN END RESULT_TIMED_OUT: [untrusted : playbooks/baremetal/collect-logs.yaml@main]

so let's add also timeout before 'oc adm' command to make sure
that it would be "killed" in proper amount of time.

Signed-off-by: Daniel Pawlik <dpawlik@redhat.com>