[test-operator] Improve variables overwriting
As part of processing test-operator stages, there is a task
that takes all defined Ansible variables and overrides them
with matching stage-specific variables. However, this task
focuses in the end only on variables specific to test-operator.
Yet, because the filtering is done on Ansible task-level,
we end up with 80% of the `ansible-post-deployment.log`
filled with entries like `skipping <variable_name>`.
To make this worse, it exposes some deployment-time secrets
as a side-effect.

This commit moves the filtering from task-level to earlier
phase ~ when the loop arguments are prepared, so we will
do override only on the test-operator-related variables,
reducing the useless output and avoiding secrets exposure.