Move cifmw_basedir var to group_vars - part four
The `cifmw_basedir` variable is now expected to be defined group_vars.
Previous fallback to `ansible_user_dir ~ '/ci-framework-data'` has been
removed from playbooks and roles.

This change ensures consistency across roles and prevents accidental use
of a default path when `cifmw_basedir` is intentionally unset.

This commit is one part of a set of changes.

Signed-off-by: Daniel Pawlik <dpawlik@redhat.com>