Update custom ca certs hook
The hook was supposed to create a secret with custom ca certs and
also update the Openstack control-plane to use it. This approach
has brought some issues with the control-plane restarting, so we
want to change the approach to creating the secret only. The use of
the secret will be done separately.

Signed-off-by: Katarina Strenkova <kstrenko@redhat.com>