[os_must_gather] Increase default timeout to 30 minutes
Earlier default timeout (10m) was not enough for some jobs and the
must gather command was not killed gracefully. After we added timeout
with addition 15 minutes, the CI jobs seems to be more stable and
the job results seems to have the logs. It is because the value
"fits" needed time, but we don't have time reservation.
Let's set 30 minutes just in case.

Signed-off-by: Daniel Pawlik <dpawlik@redhat.com>