Bump Ceph to Tentacle for upstream hci jobs
Storage operators recently switched from Prow to Zuul for both kuttl and
tempest jobs.
The new jobs are based on podified-multinode-hci-deployment-crc-1comp-backends,
where Ceph is deployed using the defaults coming from ci-framework (Reef).
However, while Reef remains the target for Antelope, we would like to bump our
CI target to Tentacle (the latest released version). For this reason this patch
proposes a set of overrides for the centos-9/hci_ceph_backends that are included
in such jobs. In addition, by adding these variables in a common place, we avoid
duplicating and maintaining the same set of parameters for several operators.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>