[libvirt_manager] Generate VMs inventory files
The libvirt_manager is adding the VMs into the inventory during the
playbook execution. It means, if Zironic would bootstrap the host,
later the CI job will continue running the CI job, but it would not be
able to reach the VMs, like controller-0.
Add a feature that dump the VMs groups into the inventory file.

Signed-off-by: Daniel Pawlik <dpawlik@redhat.com>