[devscripts] Do not remove ansible-core
By moving the bootstrap procedure from job execution to ZIronic,
we changed the way, what is configuring the baremetal host (BM).
Normally in CI job execution, Zuul is spawning a controller, then
the controller is configuring the baremetal host, then the
job continuation is done on controller-0 via baremetal host.
With Zironic, to avoid creating additional VM - controller - we
moved the procedure to be done directly on the baremetal host.
It means, that the ansible-core package needs to be available, otherwise
bootstrap procedure would fail because of missing modules, libraries
etc.

Signed-off-by: Daniel Pawlik <dpawlik@redhat.com>